Feat: 스프링 부트 버전 3.2.6으로 변경 및 초기 영화 데이터 수집 기능 구현

+++ Back_End/src/main/java/com/ssafy/youniverse/dto/res/MovieSimpleResDto.java
@@ -12,6 +12,6 @@ public class MovieSimpleResDto {
     private String title;
     private String movieImage;
     private List<KeywordResDto> keywordResDtos;
-    private Float rate;
+    private Double rate;
     private Integer runtime;
 }
